agent/requirements.txt
# Network scanning and device discovery
scapy>=2.5.0
netifaces>=0.11.0
python-nmap>=0.7.1

# API and communication
requests>=2.28.2
pyjwt>=2.6.0
websockets>=10.4
aiohttp>=3.8.4

# Data processing and security
pycryptodome>=3.17.0
cryptography>=39.0.2
pandas>=1.5.3
numpy>=1.24.2

# System utilities
psutil>=5.9.4
schedule>=1.1.0
python-dotenv>=1.0.0

# IoT protocols
paho-mqtt>=2.0.0
pymodbus>=3.0.2

# Development and build tools
pyinstaller>=6.0.0 